+++ application/views/request_order_approval/detail.php
@@ -152,7 +152,9 @@
       var col='';
           col += '<a href="#" onclick="saveQTY(\''+row.id_detail_ro+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Save</a>';
 
+      <?if($this->mdl_auth->CekAkses(array('menu_id'=>37, 'policy'=>'DELETE'))){?>  
           col += '&nbsp;&nbsp; | &nbsp;&nbsp; <a href="#" onclick="deleteData(\''+row.id_detail_ro+'\');" class="easyui-linkbutton" iconCls="icon-edit" plain="false">Delete</a>';
+      <?]?>
       return col;
     }
 
@@ -175,13 +177,15 @@
               saveQTY();
             }
           },
+        <?if($this->mdl_auth->CekAkses(array('menu_id'=>37, 'policy'=>'ADD'))){?>
           {
             iconCls:'icon-add',
             text:'Tambah Detail',
             handler:function(){
               newDetail();
             }
-          },          
+          },
+        <?}?>
           {
             iconCls:'icon-undo',
             text:'Kembali',
